openstack/requirements and ironic projects need to be updated
The openstack/requirements project leaves the "upper-constraint.txt"
in modified state after every stacking. So any update to the
project is not reflected in the /opt/stack/requirements directory.

openstack/ironic may be left in unsane state due to merge conflicts
of the patch. This leads to gate failing for subsequent run as the
new patch cannot be applied unless the directory /opt/stack/ironic is
cleaned up manually.

Change-Id: I971e80436fcb7f83f5c75c16d4b21650832c3df8